donation routes created

+++ controllers/folder/removeFolder.js
@@ -1,7 +1,29 @@
+const { cloudchannel } = require("googleapis/build/src/apis/cloudchannel");
+
 const removeFolder = async (request, reply, fastify) => {
   try {
     const { id } = request.params;
-    await fastify.Media.deleteMany({ folder: id });
+
+    const user_id = request.staff;
+    const isEligible = await fastify.getPermissions(user_id, [
+      { rule: "Gallery", priviledges: ["delete"] },
+    ]);
+
+    if (!isEligible) {
+      reply
+        .code(403)
+        .send({ msg: "You have no permission to perform this task." });
+      return;
+    }
+
+    const count = await fastify.Media.count({ folder: id });
+
+    if (count > 0) {
+      reply.code(403).send({
+        msg: "This folder contains media in it, delete all medias and try again",
+      });
+      return;
+    }
 
     await fastify.Folder.deleteOne({
       _id: id,
