+++ controllers/folder/removeMedia.js
@@ -1,6 +1,19 @@
 const removeMedia = async (request, reply, fastify) => {
   try {
     const { id } = request.params;
+
+    const user_id = request.staff;
+    const isEligible = await fastify.getPermissions(user_id, [
+      { rule: "Gallery", priviledges: ["delete"] },
+    ]);
+
+    if (!isEligible) {
+      reply
+        .code(403)
+        .send({ msg: "You have no permission to perform this task." });
+      return;
+    }
+
     await fastify.Media.deleteOne({
       _id: id,
     });
